helpers/requirements.txt
beautifulsoup4==4.10.0
mechanize==0.4.5
pymongo==3.12.0
python-dotenv==0.19.2
pytz==2021.1
requests==2.22.0
selenium==3.141.0
